requirements.txt
Flask==1.0.2
PyMySql==0.9.2
six==1.12.0
google-cloud-storage==1.13.1